added color and separate main.dart on files main.dart and calcalate_widget.dart
